create base controller
